Transfere tradutor de classic para módulo translator/*.py
